github-check-sorter: Adjust observers to avoid breaking
It seems that switching tabs in GitHub sometimes does something odd that
breaks ongoing observation of existing nodes, even `document.body`. This
doesn't seem to affect `document` itself though, so switch to observing
that.

Also, may as well then merge the sidebar sorting into the same observer,
in case the sidebar starts getting added and removed as tabs are
switched.