github-diff-tab-fix: Watch for style removal
GitHub removes our added `<style>` when switching tabs. Sigh. Watch for
that and re-add it as necessary.

Also, put the styles into the tag as a text node rather than using
`style.sheet.insertRule()`. Rules inserted with the latter method are
lost if the tag gets removed from the DOM.